requirements.txt
joblib==1.0.0
nltk==3.5
python-dateutil==2.8.1
pytz==2020.4
pymongo==3.11.3 
discord-py==1.6.0 
dnspython==2.0.0 
googletrans==3.1.0a0 